ai/freshcheck/requirements-docker.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.2
tensorflow==2.11.0
keras==2.11.0
torch
torchvision
numpy==1.24.3
opencv-python==4.10.0.84
Pillow==9.4.0
h5py==3.11.0
matplotlib==3.7.0